// src/dto/agricultural_holding/get.agricultural.holding.dto.ts
export class GetAgriculturalHoldingDto {
    agriculturalHoldingId: number;
    agriculturalHoldingName: string;
    agriculturalHoldingNumber: string;
    address: string;
    city: string;
    zipNumber: string;
    directorName: string;
    directorSurname: string;
    directorIdNum: string;
    phone: string;
    email: string;
}